feat(code): refactors inputManager to seperate from player
also removes any remaining local references and refers to it as a singleton in all cases
