+++ Assets/Scripts/Player/InputManager.cs
@@ -7,13 +7,8 @@ using UnityEngine.InputSystem.Utilities;
 public class InputManager : MonoBehaviour 
 {
     public static InputManager Instance { get; private set; }
-
     private PlayerInput playerInput;
     public PlayerInput.PlayerActions playerActions;
-    private PlayerMotor motor;
-    private PlayerLook look;
-
-    [SerializeField] private TaskManager taskManager;
 
     public enum LastInputType { KeyboardMouse, Playstation, Xbox, Gamepad }
     public LastInputType lastInputType { get; private set; }
@@ -24,18 +19,8 @@ public class InputManager : MonoBehaviour
     private void Awake()
     {
         Instance = this;
-
         playerInput = new PlayerInput();
         playerActions = playerInput.Player;
-
-        motor = GetComponent<PlayerMotor>();
-        look = GetComponent<PlayerLook>();
-
-        playerActions.Jump.performed += ctx => motor.Jump();
-        playerActions.Crouch.performed += ctx => motor.Crouch();
-        playerActions.Sprint.performed += ctx => motor.Sprint();
-        playerActions.Tasklist.performed += ctx => taskManager.ToggleTaskPad();
-        playerActions.Pause.performed += ctx => PauseManager.Instance.TogglePauseMenu();
 
         InputSystem.onAnyButtonPress.Call(OnAnyInputDetected);
         InputSystem.onEvent += OnAnyDeviceEvent;
@@ -122,21 +107,5 @@ public class InputManager : MonoBehaviour
         }
 
         return string.Empty;
-    }
-
-    private void Update()
-    {
-        motor.ProcessMove(playerActions.Move.ReadValue<Vector2>());
-        look.ProcessLook(playerActions.Look.ReadValue<Vector2>());
-    }
-
-    private void OnEnable()
-    {
-        playerActions.Enable();
-    }
-
-    private void OnDisable()
-    {
-        playerActions.Disable();
     }
 }
