+++ Assets/Scripts/Player/PlayerInputManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ccc974bfcd953e645943942436836363
